chore(auth): auth module extension
