+++ server/src/features/auth/refresh-token/refresh-token.module.ts
@@ -2,9 +2,9 @@ import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { AuthModule } from 'features/auth/auth.module';
-import { RefreshTokenEntity } from 'features/refresh-token/refresh-token.entity';
-import { RefreshTokenRepository } from 'features/refresh-token/refresh-token.repository';
-import { RefreshTokenService } from 'features/refresh-token/refresh-token.service';
+import { RefreshTokenEntity } from 'features/auth/refresh-token/refresh-token.entity';
+import { RefreshTokenRepository } from 'features/auth/refresh-token/refresh-token.repository';
+import { RefreshTokenService } from 'features/auth/refresh-token/refresh-token.service';
 
 @Module({
   imports: [
