+++ server/src/features/features.module.ts
@@ -7,7 +7,6 @@ import { CircuitModule } from 'features/circuits/circuit.module';
 import { CommentModule } from 'features/comments/comment.module';
 import { DisciplineModule } from 'features/disciplines/discipline.module';
 import { EventModule } from 'features/events/event.module';
-import { PasswordResetModule } from 'features/password-reset/password-reset.module';
 import { StatisticsModule } from 'features/statistics/statistics.module';
 import { ThreadModule } from 'features/threads/thread.module';
 import { ThreadCategoryModule } from 'features/thread-categories/thread-category.module';
@@ -18,7 +17,6 @@ import { UserModule } from 'features/user/user.module';
     CheckModule,
     UserModule,
     AuthModule,
-    PasswordResetModule,
     ArticleModule,
     ThreadModule,
     ThreadCategoryModule,
