--- server/src/features/password-reset/password-reset.module.ts
@@ -1,21 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-
-import { PasswordResetRequestEntity } from './password-reset-request.entity';
-import { PasswordResetController } from './password-reset.controller';
-import { PasswordResetRepository } from './password-reset.repository';
-import { PasswordResetService } from './password-reset.service';
-import { UserModule } from '../user/user.module';
-import { AuthModule } from '../auth/auth.module';
-
-@Module({
-  imports: [
-    TypeOrmModule.forFeature([PasswordResetRequestEntity]),
-    AuthModule,
-    UserModule,
-  ],
-  controllers: [PasswordResetController],
-  providers: [PasswordResetService, PasswordResetRepository],
-  exports: [PasswordResetService],
-})
-export class PasswordResetModule {}
